Bug 1109248: Rollup of changes previously applied to media/webrtc/trunk/webrtc rs=jesup
UltraBlame original commit: 55c60f012e8fb51715f908ac01f6df2218607938

+++ media/webrtc/trunk/webrtc/modules/video_coding/main/source/content_metrics_processing.cc
@@ -43,7 +43,7 @@ int VCMContentMetricsProcessing::Reset() {
   return VCM_OK;
 }
 
-void VCMContentMetricsProcessing::UpdateFrameRate(uint32_t frameRate) {
+void VCMContentMetricsProcessing::UpdateFrameRate(float frameRate) {
   
   recursive_avg_factor_ = static_cast<float> (1000.0f) /
       static_cast<float>(frameRate *  kQmMinIntervalMs);
